Minor edits in amplitude histogram plugin
