+++ phycontrib/views/__init__.py
@@ -1,0 +1,1 @@
+from .amplitude_histogram import *  # noqa
